Token functionality added to authorize the user

+++ controllers/contactController.js
@@ -8,7 +8,7 @@ const createContact = asyncHandler(async(req, res) => {
         return;
     }
     try {
-        const contact = await Contact.create({ name, email, phone });
+        const contact = await Contact.create({ name, email, phone,user_id: req.user.id });
         res.status(201).json(contact);
     } catch (error) {
         console.error("Error creating contact:", error);
@@ -18,7 +18,7 @@ const createContact = asyncHandler(async(req, res) => {
 
 
 const getContacts = asyncHandler(async(req,res)=>{
-    const contact = await Contact.find();
+    const contact = await Contact.find({user_id: req.user_id});
     res.status(200).json(contact);
 });
 
@@ -27,8 +27,9 @@ const getContact = asyncHandler(async(req,res)=>{
     try {
         const {id} = req.params;
         const contact = await Contact.findById(id);
-        if(!contact){
+        if(!contact || contact.user_id.toString() !== req.user.id){
             res.status(404).json({message:"Contact Not found"})
+            return
         }
         res.status(200).json(contact);
         console.log(contact)
@@ -42,12 +43,17 @@ const getContact = asyncHandler(async(req,res)=>{
 const updateContact = asyncHandler(async(req,res)=>{
     try {
         const {id} = req.params;
-        const contact = await Contact.findByIdAndUpdate(id,req.body)
+        const contact = await Contact.findById(id);
         if (!contact){
             res.status(404).json({error: "Contact not Found"});
             return;
         }
-        const updatedContact = await Contact.findById(id)
+        if (contact.user_id.toString() !== req.user.id){
+            res.status(403);
+            throw new Error("User is not authorised to Update the contact");
+        }
+        const updatedContact = await Contact.findByIdAndUpdate(id,req.body)
+        updatedContact = await Contact.findById(id);
         res.status(200).json(updatedContact)
     } catch (error) {
         console.error("Error updating the contact", error);
@@ -58,10 +64,15 @@ const updateContact = asyncHandler(async(req,res)=>{
 const deleteContact = asyncHandler(async(req,res)=>{
     try {
         const {id} = req.params;
-        const contact = await Contact.findByIdAndDelete(id);
+        const contact = await Contact.findById(id);
         if (!contact){
             res.status(404).json({message:"Contact Not found with given ID"});
         }
+        if (contact.user_id.toString() !== req.user.id){
+            res.status(403);
+            throw new Error("User is not authorized to delete a contact");
+        }
+        const deletedContact = await Contact.findByIdAndDelete(id);
         res.status(200).json({message:"Contact Deleted"})
         
     } catch (error) {
